Crate stash

Stash is a library for efficiently storing maps of keys to values when one doesn't care what the keys are but wants blazing† fast O(1) insertions, deletions, and lookups.

Common use cases include file descriptor tables, session tables, or MIO context tables.

†Blazing means an order of magnitude faster than hash maps and btree maps.

Modules

stash

unique_stash

Structs

Stash
An O(1) amortized table that reuses keys.

Tag
A versioned index into a UniqueStash.

UniqueStash
An O(1) amortized table that does not reuse keys.