Module stash::unique_stash

Structs

Extend
The iterator produced by Unique::extend.

IntoIter
Iterator over the (index, value) pairs.

IntoValues
Iterator over values in the stash.

Iter
Iterator over the (index, &value) pairs.

IterMut
Iterator over the (index, &mut value) pairs.

Tag
A versioned index into a UniqueStash.

UniqueStash
An O(1) amortized table that does not reuse keys.

Values
Iterator over references to the values in the stash.

ValuesMut
Iterator over mutable references to the values in the stash.